consumer/requirements.txt
kafka-python==2.0.2
duckdb
pandas
schedule
